pyproject.toml
[project]
name = "gcp-sensors"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "google-cloud-bigtable>=2.28.0",
    "google-cloud-core>=2.5.0",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "notebook>=7.5.3",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "google-cloud-bigquery>=3.31.0",
    "google-cloud-aiplatform>=1.30.0",
    "kfp>=2.0.0",
]
